Final Update of recipes and Edits
The last and largest update, the content of both the chicken alfredo and chicken teriyaki pages have been filled. Their format has been compared to the other recipes and made similar. There were also changtes made to grammar, spelling, etc. The links to and from home are included as well.